refactor: 💡 improve type information about nodes in Visitor
We have some issues with predicting the type of a node when traversing.
We know the type of a node only when we are visiting it, by looking up
in its type property. That is why, visitor can accept any of INode, but
the exact type will be known only when traversing in runtime. This
change requires to cast the node to specific type explicitly in visitor
implementations.