feat: 🎸 implement Value wrappers for run-time data
Makes sense to implement wrappers for a run-time data. Strings, numbers,
booleans, let us assume that we do not have an engine that handles it
for us, so we make an assumption, that we have raw binary structures
here. To easier the life for us, we need to implement wrappers for these
binary structures, so it will be easier to work with them.